requirements.txt
Django==3.1
djangorestframework==3.12
pytest==6.2
pytest-django==2.4